refactor(recordings): extract mac audio normalization into separate module
Move all macOS-specific audio normalization logic out of the shared
recording-writer.ts into a dedicated mac-audio-normalizer.ts file,
following the same pattern as mac-meeting.ts / windows-meeting.ts.

recording-writer.ts now has zero mac-specific logic in it — it imports
createChunkNormalizer from the mac module and only calls it when
process.platform === 'darwin'. On Windows/Linux, the normalizer is
never instantiated and chunks are written directly to disk.
